+++ tt-media-server/cpp_server/src/runtime/runners/llm_runner/hash.cpp
@@ -10,7 +10,7 @@ namespace tt::runners::llm_engine {
 static constexpr uint64_t K_FNV_PRIME = 1099511628211ULL;
 static constexpr uint64_t K_FNV_OFFSET = 0xcbf29ce484222325ULL;
 
-int64_t hashTokenIds(const std::vector<int64_t>& tokenIds, int64_t prefix) {
+int64_t hashTokenIds(const std::vector<uint32_t>& tokenIds, int64_t prefix) {
   uint64_t h = K_FNV_OFFSET;
   if (prefix != -1) {
     uint64_t p = static_cast<uint64_t>(prefix);
@@ -21,7 +21,7 @@ int64_t hashTokenIds(const std::vector<int64_t>& tokenIds, int64_t prefix) {
     }
   }
   const uint8_t* data = reinterpret_cast<const uint8_t*>(tokenIds.data());
-  size_t len = tokenIds.size() * sizeof(int64_t);
+  size_t len = tokenIds.size() * sizeof(uint32_t);
   for (size_t i = 0; i < len; ++i) {
     h ^= data[i];
     h *= K_FNV_PRIME;
